fix: Select window (#17)
`%idx` syntax means the index of the pane. We can list all the panes using `tmux list-panes -a`. And we can select a window that holds the pane with the given number using `tmux select-window -t %<pane_idx>`. In some cases it might work, if the id of window is equal to the id of the pane it holds. But the proper syntax we should use is `@idx`, which means the number of the window.